refactor: refactor UI
