+++ src/ui/generic/progress_bar.ts
