+++ src/ui/generic/sprites.ts
@@ -1,15 +1,18 @@
 const enum Sprites {
   DELETE_DATA = 5178,
+
   SYNC_PAUSED = 5179,
   SYNC_RUNNING = 5180,
   SYNC_MANUAL = 5181,
   SYNC_RELOAD = 29445,
+
   INDICATOR_PAUSED_UNLIT = 29374,
   INDICATOR_PAUSED_LIT = INDICATOR_PAUSED_UNLIT + 2,
   INDICATOR_MANUAL_UNLIT = 29378,
   INDICATOR_MANUAL_LIT = INDICATOR_MANUAL_UNLIT + 2,
   INDICATOR_RUNNING_UNLIT = 29382,
   INDICATOR_RUNNING_LIT = INDICATOR_RUNNING_UNLIT + 2,
+
   SHOW_PROGRESS_BAR = 29394,
   SHOW_VALUE = 29395
 }
