+++ src/ui/ui.ts
@@ -1,7 +1,7 @@
 import Environment from "../common/environment"
 import Logger from "../utils/logger"
-import { initMainMenu, openMainMenu } from "./main"
-import { initToolboxMenu } from "./toolbox"
+import * as Main from "./main"
+import * as Toolbox from "./toolbox"
 
 namespace UI {
   /**
@@ -11,14 +11,14 @@ namespace UI {
    */
   export function init(): void {
     Logger.assert(Environment.isUiAvailable, "UI is not available.")
-    initMainMenu()
-    initToolboxMenu()
+    Main.init()
+    Toolbox.init()
     ui.registerShortcut({
       id: "dynamicdashboard.main.open",
       text: "Open Dynamic Dashboard",
       bindings: ["SHIFT+D"],
       callback: () => {
-        openMainMenu()
+        Main.open()
       }
     })
   }
